+++ packages/liveblocks-react-lexical/src/liveblocks-plugin-provider.tsx
@@ -36,7 +36,7 @@ export function useIsEditorReady(): boolean {
 
   const subscribe = useCallback(
     (callback: () => void) => {
-      if (yjsProvider === undefined) return () => {};
+      if (yjsProvider === undefined) return () => { };
       yjsProvider.on("status", callback);
       return () => {
         yjsProvider.off("status", callback);
@@ -141,8 +141,9 @@ export const LiveblocksPlugin = ({
 
   const providerFactory = useCallback(
     (id: string, yjsDocMap: Map<string, Doc>): Provider => {
-      const provider = getYjsProviderForRoom(room);
+      const provider = getYjsProviderForRoom(room, {}, true);
       yjsDocMap.set(id, provider.getYDoc());
+
 
       return provider as Provider;
     },