+++ packages/liveblocks-react/src/index.ts
@@ -82,6 +82,7 @@ export {
   useSelf,
   useStorage,
   useThreads,
+  useSearchComments,
   useAttachmentUrl,
   useHistoryVersions,
   useRoomSubscriptionSettings,